Extract filters and modals into components
